added a 2 locations display (with no route highlighted)
